requirements.txt
absl-py==2.1.0
altair==5.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
attrs==24.1.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.1
cycler==0.12.1
deap==1.4.1
debugpy @ file:///D:/bld/debugpy_1707444544473/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepface==0.0.92
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.8.0.post1
feature-engine==1.8.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.4
gunicorn==22.0.0
h5py==3.11.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython @ file:///D:/bld/ipython_1717182928131/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257313664/work
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.3
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.18.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
patsy==0.5.6
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.25.4
psutil @ file:///D:/bld/psutil_1705722591889/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq @ file:///D:/bld/pyzmq_1715024524369/work
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rpds-py==0.19.1
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.13.1
seaborn==0.13.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
soupsieve==2.5
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.2
stopit==1.1.2
streamlit==1.37.0
streamlit-cropper==0.2.2
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.7.24
toml==0.10.2
toolz==0.12.1
torch==2.0.0+cpu
torchaudio==2.0.0+cpu
torchvision==0.15.0+cpu
tornado @ file:///D:/bld/tornado_1708363244199/work
TPOT==0.12.2
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
ultralytics==8.2.72
ultralytics-thop==2.0.0
update-checker==0.18.0
urllib3==2.2.1
watchdog==4.0.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.0.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
